requirements.txt
Django==1.10.2
djangorestframework==3.4.7
redis==2.10.5
tornado==4.4.2
